+++ classes/Organisations.js
@@ -2,6 +2,7 @@
 'use strict'
 
 const OneBlinkAPI = require('../lib/one-blink-api.js')
+const uploadAsset = require('../lib/upload-asset.js')
 
 module.exports = (tenant /* : Tenant */) =>
   class Organisations extends OneBlinkAPI {
@@ -20,5 +21,43 @@ module.exports = (tenant /* : Tenant */) =>
       }
 
       return super.getRequest(`/organisations/${organisationId}`)
+    }
+
+    uploadAsset(
+      assetData /* :string */,
+      assetFileName /*:string */,
+    ) /* :Promise<{location :string}> */ {
+      if (typeof assetData !== 'string') {
+        return Promise.reject(
+          new TypeError('Must supply "assetData" as a string'),
+        )
+      }
+      if (typeof assetFileName !== 'string') {
+        return Promise.reject(
+          new TypeError('Must supply "assetFileName" as a string'),
+        )
+      }
+
+      return super
+        .getRequest('/organisations')
+        .then((searchResponse) => {
+          if (!searchResponse.organisations || !searchResponse.organisations[0]) {
+            return Promise.reject(
+              new TypeError('You do not have access to any organisations'),
+            )
+          }
+          return searchResponse.organisations[0].id
+        })
+        .then((organisationId) => {
+          return super.postRequest('/asset-upload-credentials', {
+            assetPath: `assets/${assetFileName}`,
+            organisationId,
+          })
+        })
+
+        .then((credentials) => uploadAsset(credentials, assetData))
+        .then((uploadDetails) => {
+          return { location: uploadDetails.Location }
+        })
     }
   }